Split messageFilter method into trait
