+++ src/Target.php
@@ -4,8 +4,6 @@ declare(strict_types=1);
 
 namespace Yiisoft\Profiler;
 
-use Yiisoft\Strings\WildcardPattern;
-
 /**
  * Target is the base class for all profiling target classes.
  *
@@ -17,33 +15,11 @@ use Yiisoft\Strings\WildcardPattern;
  */
 abstract class Target
 {
+    use MessageFilterTrait;
     /**
      * @var bool whether to enable this log target. Defaults to true.
      */
     public bool $enabled = true;
-
-    /**
-     * @var array list of message categories that this target is interested in. Defaults to empty, meaning all
-     * categories.
-     *
-     * You can use an asterisk at the end of a category so that the category may be used to
-     * match those categories sharing the same common prefix. For example, 'Yiisoft\Db\*' will match
-     * categories starting with 'Yiisoft\Db\', such as `Yiisoft\Db\Connection`.
-     */
-    public array $categories = [];
-
-    /**
-     * @var array list of message categories that this target is NOT interested in. Defaults to empty, meaning no
-     * uninteresting messages.
-     *
-     * If this property is not empty, then any category listed here will be excluded from {@see categories}.
-     * You can use an asterisk at the end of a category so that the category can be used to
-     * match those categories sharing the same common prefix. For example, 'Yiisoft\Db\*' will match
-     * categories starting with 'Yiisoft\Db\', such as `Yiisoft\Db\Connection`.
-     *
-     * {@see categories}
-     */
-    public array $except = [];
 
     /**
      * Processes the given log messages.
@@ -75,40 +51,4 @@ abstract class Target
      * @param Message[] $messages profiling messages to be exported.
      */
     abstract public function export(array $messages);
-
-    /**
-     * Filters the given messages according to their categories.
-     *
-     * @param Message[] $messages messages to be filtered.
-     * The message structure follows that in {@see Profiler::$messages}.
-     *
-     * @return array the filtered messages.
-     */
-    protected function filterMessages(array $messages): array
-    {
-        foreach ($messages as $i => $message) {
-            $matched = empty($this->categories);
-
-            foreach ($this->categories as $category) {
-                if ((new WildcardPattern($category))->match($message->level())) {
-                    $matched = true;
-                    break;
-                }
-            }
-
-            if ($matched) {
-                foreach ($this->except as $category) {
-                    if ((new WildcardPattern($category))->match($message->level())) {
-                        $matched = false;
-                        break;
-                    }
-                }
-            }
-
-            if (!$matched) {
-                unset($messages[$i]);
-            }
-        }
-        return $messages;
-    }
 }
